src/requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.7.0rc1
Django==3.1.4
django-cors-headers==3.5.0
django-environ==0.4.5
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.8
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
pyparsing==2.4.7
pytest==6.2.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.26.2
zope.interface==5.2.0
